Refactor theme utilities by moving getIsDarkMode to a new themeUtils file for better organization. Update imports in useTheme and themeColors to reflect the new location, enhancing code clarity and maintainability.
